FEM core functions (required in PASSMO) moved to femutils for general FEM use.
